More Database Implementation
Updates to the database. Specifically, I am adding the day and month of the workout to the database. This way, when the user click on a specific day, it will only list the workouts on the specific day and month. I am passing the day and month over multiple files to make sure that the correct day and month are added to the database. I updated the WorkoutDayList fragment so that if the list is empty on that day, the workoutdayempty fragment will replace workoutdaylist. The user can then add workouts for that day.